requirements.txt
# Web framework for API
fastapi==0.101.1
uvicorn[standard]==0.23.2

# Google Cloud Storage access
google-cloud-storage==2.12.0
google-cloud-speech==3.3.3
google-cloud-texttospeech==2.16.1

# OpenAI API for embeddings & LLM
openai==1.8.4

# Vector store (ChromaDB) for storing embeddings
chromadb==0.3.36

# Environment variable loading from .env
python-dotenv==1.0.0

# For Streamlit UI
streamlit==1.33.0
streamlit-mic-recorder==0.0.4
requests==2.31.0

# Agentic Framework
crewai
langchain
